refactored test run card code structure
