+++ client/app/components/blocks/blocks.js
@@ -1,3 +1,2 @@
-require('./test-run-card/test-run-card.module');
 require('./app-sidebar/app-sidebar.module');
 require('./app-header/app-header.module');
