--- client/app/components/blocks/test-run-card/test-run-card.module.js
@@ -1,7 +1,0 @@
-(function () {
-    'use strict';
-
-    angular.module('app.testRunCard', []);
-
-    require('./test-run-card.directive');
-})();
