+++ client/app/shared/test-run-card/test-run-card.module.js
@@ -1,0 +1,7 @@
+'use strict';
+
+import testRunCard from './test-run-card.directive';
+
+export const testRunCardModule = angular.module('testRunCard', [])
+    .directive({ testRunCard })
+    .name;
